+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2
+
+* saveFileDialog now directly write content to the selected file, and returns the name of the created file
+* openFileDialog now directly returns the content of the opened file
+* removed folderDialog function
+
 ## 0.1.0
 
 * Initial release : you can show a dialog in order to open a file, in order to save a file or in order to select a folder.